Upgrade winmd generation to ClangSharp 21

Copilot-Session: 19e5498a-dfe7-4601-8955-98f33a92582b

+++ generation/WinSDK/RecompiledIdlHeaders/um/dxva9typ.h
@@ -261,9 +261,9 @@ typedef enum _DXVA_SampleFormat {
 #define DXVA_ExtractExtColorData(_sf, _Mask, _Shift) \
     (((_sf) & (_Mask)) >> (_Shift))
 
-#define DXVABitMask(__n) (~((~0) << __n))
+#define DXVABitMask(__n) (~((~0u) << __n))
 #define DXVA_ExtColorData_ShiftBase 8
-#define DXVAColorMask(__bits,__base) (DXVABitMask(__bits) << (__base))
+#define DXVAColorMask(__bits,__base) ((int)(DXVABitMask(__bits) << (__base)))
 
 typedef enum _DXVA_VideoTransferFunction
 {
@@ -1005,5 +1005,4 @@ typedef struct _DXVA_COPPStatusSignalingCmdData {
 
 #endif /* WINAPI_FAMILY_PARTITION(WINAPI_PARTITION_DESKTOP) */
 #pragma endregion
-
